ウィングドーザNPCを新形式にした
